cli: move the rewrite entrypoint to the module root; rename engine->single
The dispatcher (run/run_with_writer) logically sits above both pipelines but was
living inside the single-input file and reaching sideways into merge. Move it to
the rewrite module root and rename engine.rs to single.rs so the layering reads:
root dispatch -> single / merge, both on common + options.

Pure reorganization; no behavior change.
